@@ -39,27 +39,28 @@
     "LICENCE",
     "*.md"
   ],
-  "dependencies": {
-    "@angular/core": "^5.2.0",
-    "@angular/http": "^5.2.0"
+  "peerDependencies": {
+    "@angular/core": ">=5.0.0 <7.0.0",
+    "@angular/http": ">=5.0.0 <7.0.0"
   },
   "devDependencies": {
-    "@angular/cli": "^1.7.2",
-    "@angular/common": "^5.2.0",
-    "@angular/compiler": "^5.2.0",
-    "@angular/compiler-cli": "^5.2.0",
-    "@angular/core": "^5.2.0",
-    "@angular/forms": "5.2.0",
-    "@angular/platform-browser": "^5.2.0",
-    "@angular/platform-browser-dynamic": "^5.2.0",
-    "@angular/router": "^5.2.0",
+    "@angular/cli": "^6.0.0",
+    "@angular/common": "^6.0.0",
+    "@angular/compiler": "^6.0.0",
+    "@angular/compiler-cli": "^6.0.0",
+    "@angular/core": "^6.0.0",
+    "@angular/forms": "^6.0.0",
+    "@angular/http": "^6.0.0",
+    "@angular/platform-browser": "^6.0.0",
+    "@angular/platform-browser-dynamic": "^6.0.0",
+    "@angular/router": "^6.0.0",
     "@types/core-js": "^0.9.41",
     "@types/jasmine": "2.5.36",
     "@types/node": "~6.0.60",
     "angular2-template-loader": "^0.6.2",
     "bootstrap": "^3.3.7",
     "bootstrap-material-design": "^0.5.10",
-    "codelyzer": "~2.0.0",
+    "codelyzer": "~4.2.1",
     "conventional-changelog-cli": "^1.3.1",
     "conventional-github-releaser": "^1.1.3",
     "coveralls": "^2.13.1",
@@ -79,12 +80,13 @@
     "npm-run-all": "^4.0.2",
     "protractor": "~5.1.0",
     "reflect-metadata": "^0.1.10",
-    "rxjs": "^5.5.6",
-    "ts-node": "~4.1.0",
+    "rxjs": "^6.0.0",
+    "ts-node": "~5.0.1",
     "tslint": "~5.9.1",
-    "typescript": "~2.5.3",
+    "typescript": "~2.7.2",
     "wallaby-webpack": "0.0.38",
     "web-animations-js": "^2.2.5",
-    "zone.js": "^0.8.19"
+    "zone.js": "^0.8.4",
+    "@angular-devkit/build-angular": "~0.6.0"
   }
 }
