@@ -1,5 +1,5 @@
 import {Injectable} from '@angular/core';
-import {Observable} from 'rxjs/Observable';
+import {Observable} from 'rxjs';
 import {NgxOAuthConfig} from './config-interface';
 import {HttpClient, HttpHeaders, HttpParams} from '@angular/common/http';
 import {NgxRequest} from './ngx-request';
